Removed checking of specific heroes of the storm temp path
Watching for this directory is sketchy as it doesnt exist.  Previous iteration had the tool creating it, but then it sometimes was deleted while watching or wasnt creating correctly for watching.

Watching the entire temp directory presents no issues as long as we specify whast file type to watch for, which we are.  Need to fully test performance impacts though